requirements.txt
# 基础依赖
numpy==1.24.3
scipy==1.11.3
pandas==2.0.3
matplotlib==3.7.2
scikit-learn==1.3.0
tqdm==4.66.1

# 音频处理
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
webrtcvad==2.0.10
noisereduce==2.0.1
audiomentations==0.30.0
praat-parselmouth==0.4.3

# 深度学习
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
transformers==4.35.0
timm==0.9.5
efficientnet-pytorch==0.7.1
pytorch-lightning==2.1.0

# 图像处理
Pillow==10.0.0
opencv-python==4.8.1.78

# 可视化和评估
seaborn==0.13.0
tensorboard==2.14.1
scikit-plot==0.3.7

# 其他工具
pyaudio==0.2.13
pyyaml==6.0.1
argparse>=1.4.0
joblib==1.3.2 